Momentum offset for 6d orbit
